move fake definition and fake schema logic to a fakeDefintions directory + update imports
