+++ src/utils/graphql.ts
@@ -5,8 +5,8 @@ import {
   buildClientSchema,
 } from 'graphql';
 import { Options } from 'express-graphql';
-// TODO - fake_schema kept seperate, it should potentially be moved out of the utils folder
-import { fakeFieldResolver, fakeTypeResolver } from 'src/utils/fake_schema';
+
+import { fakeFieldResolver, fakeTypeResolver } from 'src/fakeDefinitions';
 import { getProxyExecuteFn } from 'src/utils';
 
 export type Maybe<T> = T | null | undefined;
